+++ src/main/java/com/telepathicgrunt/world_blender/WorldBlenderClient.java
@@ -5,7 +5,6 @@ import com.telepathicgrunt.world_blender.blocks.WBPortalBlockEntity;
 import com.telepathicgrunt.world_blender.blocks.WBPortalBlockEntityRenderer;
 import com.telepathicgrunt.world_blender.dimension.WBSkyProperty;
 import com.telepathicgrunt.world_blender.mixin.SkyPropertiesAccessor;
-import com.telepathicgrunt.world_blender.utils.GoVote;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendereregistry.v1.BlockEntityRendererRegistry;
 import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
@@ -16,7 +15,6 @@ import net.minecraft.util.math.BlockPos;
 public class WorldBlenderClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
-		GoVote.init();
 		SkyPropertiesAccessor.getBY_IDENTIFIER().put(new Identifier(WorldBlender.MODID, "sky_property"), new WBSkyProperty());
 		BlockEntityRendererRegistry.INSTANCE.register(WBBlocks.WORLD_BLENDER_PORTAL_BE, WBPortalBlockEntityRenderer::new);